+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Table size limit in `MarkdownUtils.formatDataTable(List, int)` method, by @HardNorth
+- Transpose table logic if column number is bigger than row number in `MarkdownUtils.formatDataTable(List, int)` method, by @HardNorth
 - `MarkdownUtils.formatDataTable(Map<String, String>)` method, by @HardNorth
 
 ## [5.2.1]